+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 * fixed password issue when updating the BackendlessUser object
 * fixed Date type issue when retrieving object or registering BackendlessUser with custom Date field
 * fixed the BackendlessUser properties methods
+* added support of custom smart-text substitutions for push templates:
+```
+func pushWithTemplate(templateName: String, templateValues: [String : Any], responseHandler: ((MessageStatus) -> Void)!, errorHandler: ((Fault) -> Void)!)
+```
 
 ### 0.0.9, / June, 6 2019
 * added IEmailEnvelope protocol, EmailEnvelope, EnvelopeWithRecepients and EnvelopeWithQuery classes